metrics.h: changed time point type to long long int nanoseconds from double, as sub-nanosecond precision unlikely to be accurate or helpful...

+++ src/metrics.h
@@ -7,7 +7,7 @@
 #include <map>
 #include <chrono>
 
-using ns = std::chrono::duration<double,std::nano>;
+using ns = std::chrono::duration<long long int,std::nano>;
 
 void metricsInit(void);
